feat(front): add front controller to project
